BackEnd/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic>=2.5.3
pydantic-settings>=2.10.1
httpx>=0.27.0,<0.28.0
pytest==8.0.0
pytest-asyncio
python-dotenv>=1.0.1
python-jose[cryptography]==3.3.0
langchain
langchain-community
langchain-experimental
langchain-ollama
langchain-core
qdrant-client
langchain-qdrant
pypdf
email-validator
passlib==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
mlflow>=2.14.0
scikit-learn==1.4.0
pandas==2.2.0
deepeval
prometheus-fastapi-instrumentator==7.0.0
langchain-openai>=0.1.0
asyncio